+++ R/ac_utils.R
@@ -1,0 +1,6 @@
+ac_progress_bar_set <- function() {
+
+  options(cli.progress_show_after = 0)
+  options(cli.progress_clear = FALSE)
+
+}